generated_code/batch_processing.py
import schedule
import time

from data_pipeline import DataPipeline
from data_validation import validate_data

# Configure data pipeline and data sources
data_sources = [
    FileDataSource("path/to/file.csv"),
    ApiDataSource("https://api.example.com/data"),
]
pipeline = DataPipeline(data_sources)

def process_data():
    """Executes the data pipeline and validation."""
    data = pipeline.run()
    validated_data = validate_data(data)
    # Perform batch processing on validated_data
    # Example: write to database, generate reports, etc.

# Schedule batch processing job
schedule.every().day.at("10:00").do(process_data)

while True:
    schedule.run_pending()
    time.sleep(1)