generated_code/real_time_processing.py
from flask import Flask, request, jsonify

from data_pipeline import DataPipeline
from data_validation import validate_data

app = Flask(__name__)

# Configure data pipeline and data sources (for real-time updates)

@app.route('/process_data', methods=['POST'])
def process_data_route():
    """Endpoint to handle real-time data processing."""
    data = request.get_json()
    pipeline = DataPipeline([ApiDataSource("https://api.realtime.com/data")])
    data = pipeline.run()
    validated_data = validate_data(data)
    # Perform real-time processing on validated_data
    # Example: update dashboards, trigger alerts, etc.
    return jsonify({'status': 'success'})

if __name__ == '__main__':
    app.run(debug=True)