requirements.txt
# requirements.txt
# AI SDLC Wizard - Professional Edition
# Complete dependency list with version pinning for stability

# Core Framework
streamlit==1.45.0
streamlit-extras==0.5.0
streamlit-option-menu==0.3.13

# LangGraph and LangChain
langgraph==0.3.5
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langchain==0.3.20
langchain-core==0.3.59
langchain-groq==0.2.5
langchain-text-splitters==0.3.8
langsmith==0.3.42

# AI and ML Models
groq==0.18.0
pydantic==2.10.6
pydantic_core==2.27.2

# Data Processing
pandas==2.2.3
numpy==2.2.5
pyarrow==20.0.0

# Visualization and Charts
plotly==5.22.0
kaleido==0.2.1
altair==5.5.0

# Document Generation
python-docx==1.1.2
reportlab==4.2.5
openpyxl==3.1.5
python-pptx==1.0.2

# File Processing
python-multipart==0.0.20
lxml==5.4.0
PyYAML==6.0.2

# Environment and Configuration
python-dotenv==1.0.1
click==8.2.0
toml==0.10.2

# Networking and APIs
requests==2.32.3
requests-toolbelt==1.0.0
httpx==0.28.1
httpcore==1.0.9
fastapi==0.115.8
uvicorn==0.34.0
starlette==0.45.3

# Database
SQLAlchemy==2.0.40

# Caching and Storage
diskcache==5.6.3
cachetools==5.5.2

# Monitoring and Logging
psutil==6.1.1
rich==13.9.4
watchdog==6.0.0

# Development and Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.10.0
flake8==7.1.1
pylint==3.3.3
bandit==1.8.0

# Utilities
humanize==4.11.0
tabulate==0.9.0
tenacity==9.1.2

# Core Python Libraries
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
six==1.17.0

# JSON and Data Formats
orjson==3.10.18
ormsgpack==1.9.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1

# Template and Markup
Jinja2==3.1.6
MarkupSafe==3.0.2

# Async and Concurrency
anyio==4.9.0
sniffio==1.3.1

# System and OS
packaging==24.2
distro==1.9.0
tornado==6.4.2

# Protobuf and Serialization
protobuf==6.30.2

# Git Integration (for version control features)
GitPython==3.1.44
gitdb==4.0.12
smmap==5.0.2

# Additional Utilities
attrs==25.3.0
annotated-types==0.7.0
blinker==1.9.0
referencing==0.36.2
rpds-py==0.24.0
typing_extensions==4.13.2
narwhals==1.39.0

# Date and Time
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Compression
zstandard==0.23.0

# Image Processing (for future features)
pillow==11.2.1

# Streamlit Specific
pydeck==0.9.1

# Platform Specific (for advanced features)
h11==0.16.0

# Optional: Enhanced UI Components (uncomment if needed)
# streamlit-aggrid==1.0.5
# streamlit-elements==0.1.0

# Optional: Additional AI Models (uncomment if needed)
# transformers==4.35.0
# torch==2.1.0
# tensorflow==2.14.0

# Optional: Database Drivers (uncomment as needed)
# psycopg2-binary==2.9.7  # PostgreSQL
# pymongo==4.5.0          # MongoDB
# redis==5.0.0            # Redis
# elasticsearch==8.10.0   # Elasticsearch

# Optional: Cloud Services (uncomment as needed)
# boto3==1.28.62          # AWS
# google-cloud-storage==2.10.0  # Google Cloud
# azure-storage-blob==12.18.3   # Azure

# Development only (install with: pip install -r requirements-dev.txt)
# Create requirements-dev.txt with additional dev dependencies:
# jupyterlab==4.0.7
# ipython==8.16.1
# pre-commit==3.4.0
# mypy==1.6.1